注意 service 的服务器名称,和客户端同步,不然返回值就0,按道理应该有个错误码?

+++ learnTopic_ws/src/server/src/main.cpp
@@ -8,8 +8,8 @@
 bool callback(roscpp_tutorials::TwoInts::Request &request,roscpp_tutorials::TwoInts::Response& response)
 {
     /** request */
-    int a = request.a;
-    int b = request.b;
+    long a = request.a;
+    long b = request.b;
     /** response */
     response.sum = a + b;
